# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: LaunchFlow multi-tenant SaaS platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /public/{tenantSlug}/site:
    get:
      operationId: getPublicSite
      tags: [tenants]
      summary: Get public site data for a tenant (homepage, settings, etc.)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public site data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSiteData"

  # ─── Leads ───────────────────────────────────────────────────────────────
...
  /projects:
    get:
      operationId: listProjects
      tags: [projects]
      summary: List projects for current tenant
      parameters:
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      operationId: createProject
      tags: [projects]
      summary: Create a project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Created project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
  /projects/{id}:
    get:
      operationId: getProject
      tags: [projects]
      summary: Get project by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    patch:
      operationId: updateProject
      tags: [projects]
      summary: Update project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdate"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
    delete:
      operationId: deleteProject
      tags: [projects]
      summary: Delete project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /projects/{id}/updates:
    get:
      operationId: listProjectUpdates
      tags: [projects]
      summary: List updates for a project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Project updates
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProjectUpdate"
    post:
      operationId: createProjectUpdate
      tags: [projects]
      summary: Add update to project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectUpdateInput"
      responses:
        "201":
          description: Created update
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectUpdate"

  # ─── Customers ────────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/services:
    get:
      operationId: listPublicServices
      tags: [services]
      summary: List published services for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published services
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Service"
  /public/{tenantSlug}/services/{slug}:
    get:
      operationId: getPublicService
      tags: [services]
      summary: Get a published service by slug (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Service
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Service"

  # ─── Area Pages ───────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/areas:
    get:
      operationId: listPublicAreas
      tags: [areas]
      summary: List published area pages for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Published area pages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AreaPage"
  /public/{tenantSlug}/areas/{slug}:
    get:
      operationId: getPublicArea
      tags: [areas]
      summary: Get a published area page by slug (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Area page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AreaPage"

  # ─── Gallery ──────────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/gallery:
    get:
      operationId: browsePublicGallery
      tags: [gallery]
      summary: List gallery images for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Gallery images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryImage"

  # ─── Before & After ───────────────────────────────────────────────────────
...
  /public/{tenantSlug}/before-after:
    get:
      operationId: listPublicBeforeAfter
      tags: [before-after]
      summary: List before/after transformations for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Transformations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BeforeAfter"

  # ─── Reviews ──────────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/reviews:
    get:
      operationId: listPublicReviews
      tags: [reviews]
      summary: List reviews for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Reviews
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Review"

  # ─── Case Studies ─────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/case-studies:
    get:
      operationId: listPublicCaseStudies
      tags: [case-studies]
      summary: List published case studies for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case studies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CaseStudy"
  /public/{tenantSlug}/case-studies/{slug}:
    get:
      operationId: getPublicCaseStudy
      tags: [case-studies]
      summary: Get published case study by slug (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Case study
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseStudy"

  # ─── FAQs ─────────────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/faqs:
    get:
      operationId: listPublicFaqs
      tags: [faqs]
      summary: List FAQs for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: FAQs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FAQ"

  # ─── Blog ─────────────────────────────────────────────────────────────────
...
  /public/{tenantSlug}/blog:
    get:
      operationId: browsePublicBlog
      tags: [blog]
      summary: List published blog posts for a tenant (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlogPost"
  /public/{tenantSlug}/blog/{slug}:
    get:
      operationId: getPublicBlogPost
      tags: [blog]
      summary: Get published blog post by slug (public)
      parameters:
        - name: tenantSlug
          in: path
          required: true
          schema:
            type: string
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Blog post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlogPost"

  # ─── Visualiser Requests ──────────────────────────────────────────────────
...
components:
  schemas:
...
    Tenant:
      type: object
      required: [id, name, slug, businessType, status]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        businessType:
          type: string
        domain:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        adminEmail:
          type: ["string", "null"]
        address:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        primaryColor:
          type: ["string", "null"]
        secondaryColor:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, suspended]
        plan:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    PublicSiteData:
      type: object
      required: [tenant]
      properties:
        tenant:
          $ref: "#/components/schemas/Tenant"
        settings:
          $ref: "#/components/schemas/TenantSettings"
        featuredServices:
          type: array
          items:
            $ref: "#/components/schemas/Service"
        featuredReviews:
          type: array
          items:
            $ref: "#/components/schemas/Review"
        recentCaseStudies:
          type: array
          items:
            $ref: "#/components/schemas/CaseStudy"
        featuredBeforeAfter:
          type: array
          items:
            $ref: "#/components/schemas/BeforeAfter"
        globalFaqs:
          type: array
          items:
            $ref: "#/components/schemas/FAQ"

    # ─── Tenant Settings ──────────────────────────────────────────────────
    TenantSettings:
      type: object
      required: [id, tenantId]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        footerContent:
          type: ["string", "null"]
        privacyPolicy:
          type: ["string", "null"]
        termsContent:
          type: ["string", "null"]
        cookiePolicy:
          type: ["string", "null"]
        openingHours:
          type: ["string", "null"]
        facebookUrl:
          type: ["string", "null"]
        instagramUrl:
          type: ["string", "null"]
        twitterUrl:
          type: ["string", "null"]
        linkedinUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        tiktokUrl:
          type: ["string", "null"]
        serviceAreas:
          type: array
          items:
            type: string
...
    Project:
      type: object
      required: [id, tenantId, title, status, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        quoteId:
          type: ["integer", "null"]
        customerId:
          type: ["integer", "null"]
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [Enquiry, "Survey Booked", "Quote Approved", Scheduled, "In Progress", Completed]
        startDate:
          type: ["string", "null"]
        endDate:
          type: ["string", "null"]
        warrantyInfo:
          type: ["string", "null"]
        assignedStaffIds:
          type: array
          items:
            type: integer
        imageUrls:
          type: array
          items:
            type: string
        documentUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ProjectInput:
      type: object
      required: [title]
      properties:
        quoteId:
          type: integer
        customerId:
          type: integer
        title:
          type: string
          minLength: 1
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        warrantyInfo:
          type: string
        assignedStaffIds:
          type: array
          items:
            type: integer
    ProjectUpdate:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        title:
          type: string
        description:
          type: string
        status:
          type: string
        startDate:
          type: string
        endDate:
          type: string
        warrantyInfo:
          type: string
        assignedStaffIds:
          type: array
          items:
            type: integer
        imageUrls:
          type: array
          items:
            type: string
        documentUrls:
          type: array
          items:
            type: string
        content:
          type: string
        isCustomerVisible:
          type: boolean
        authorId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
    ProjectUpdateInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
          minLength: 1
        isCustomerVisible:
          type: boolean

    # ─── Customer ─────────────────────────────────────────────────────────
...
    Service:
      type: object
      required: [id, tenantId, name, slug, published]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        slug:
          type: string
        published:
          type: boolean
        featuredImageUrl:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        heroContent:
          type: ["string", "null"]
        mainContent:
          type: ["string", "null"]
        benefits:
          type: array
          items:
            type: string
        processSteps:
          type: array
          items:
            type: string
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    AreaPage:
      type: object
      required: [id, tenantId, name, slug, published]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        name:
          type: string
        slug:
          type: string
        published:
          type: boolean
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        heroTitle:
          type: ["string", "null"]
        introContent:
          type: ["string", "null"]
        mapEmbed:
          type: ["string", "null"]
        availableServices:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
...
    GalleryImage:
      type: object
      required: [id, tenantId, imageUrl, featured, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        imageUrl:
          type: string
        title:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        areaId:
          type: ["integer", "null"]
        propertyType:
          type: ["string", "null"]
        finishType:
          type: ["string", "null"]
        featured:
          type: boolean
        createdAt:
          type: string
          format: date-time
...
    BeforeAfter:
      type: object
      required: [id, tenantId, title, beforeImageUrl, afterImageUrl, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        location:
          type: ["string", "null"]
        propertyType:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        beforeImageUrl:
          type: string
        afterImageUrl:
          type: string
        description:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    Review:
      type: object
      required: [id, tenantId, customerName, rating, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        customerName:
          type: string
        location:
          type: ["string", "null"]
        rating:
          type: integer
          minimum: 1
          maximum: 5
        content:
          type: string
        serviceId:
          type: ["integer", "null"]
        projectId:
          type: ["integer", "null"]
        featured:
          type: boolean
        source:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...
    CaseStudy:
      type: object
      required: [id, tenantId, title, slug, published, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        slug:
          type: string
        published:
          type: boolean
        location:
          type: ["string", "null"]
        propertyType:
          type: ["string", "null"]
        serviceType:
          type: ["string", "null"]
        problem:
          type: ["string", "null"]
        solution:
          type: ["string", "null"]
        renderSystem:
          type: ["string", "null"]
        colourFinish:
          type: ["string", "null"]
        timeTaken:
          type: ["string", "null"]
        beforePhotos:
          type: array
          items:
            type: string
        afterPhotos:
          type: array
          items:
            type: string
        customerReview:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        serviceId:
          type: ["integer", "null"]
        areaId:
          type: ["integer", "null"]
        createdAt:
          type: string
          format: date-time
...
    FAQ:
      type: object
      required: [id, tenantId, question, answer, isGlobal]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        question:
          type: string
        answer:
          type: string
        isGlobal:
          type: boolean
        serviceId:
          type: ["integer", "null"]
        areaId:
          type: ["integer", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
...
    BlogPost:
      type: object
      required: [id, tenantId, title, slug, published, createdAt]
      properties:
        id:
          type: integer
        tenantId:
          type: integer
        title:
          type: string
        slug:
          type: string
        published:
          type: boolean
        content:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        featuredImageUrl:
          type: ["string", "null"]
        seoTitle:
          type: ["string", "null"]
        seoDescription:
          type: ["string", "null"]
        categoryId:
          type: ["integer", "null"]
        categoryName:
          type: ["string", "null"]
        publishedAt:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
...