# lib/api-zod/src/generated/types/service.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * LaunchFlow multi-tenant SaaS platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Service {
  id: number;
  tenantId: number;
  name: string;
  slug: string;
  published: boolean;
  /** @nullable */
  featuredImageUrl?: string | null;
  /** @nullable */
  seoTitle?: string | null;
  /** @nullable */
  seoDescription?: string | null;
  /** @nullable */
  heroContent?: string | null;
  /** @nullable */
  mainContent?: string | null;
  benefits?: string[];
  processSteps?: string[];
  sortOrder?: number;
  createdAt?: Date;
}